requirements.txt
absl-py==2.1.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
h5py==3.12.1
Jinja2==3.1.5
keras==3.8.0
numpy==2.0.2
pillow==11.1.0
protobuf==5.29.3
grpcio==1.69.0
SQLAlchemy==2.0.37
tensorflow-cpu==2.18.0
Werkzeug==3.1.3
PyGithub==2.5.0
